+++ src/Baballonia.Tests/Baballonia.Tests.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>net8.0</TargetFrameworks>
-    
+
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
@@ -12,10 +12,18 @@
     <PackageReference Include="Moq" Version="4.20.72" />
     <PackageReference Include="MSTest.TestAdapter" Version="3.10.0" />
     <PackageReference Include="MSTest.TestFramework" Version="3.10.0" />
+
+      <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.8.0.20230708" Condition="$([MSBuild]::IsOSPlatform('Windows'))" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Baballonia\Baballonia.csproj" />
   </ItemGroup>
-  
+
+    <ItemGroup>
+        <Reference Include="OpenCvSharp">
+            <HintPath>..\..\..\opencvsharp\src\OpenCvSharp\bin\Debug\net6.0\OpenCvSharp.dll</HintPath>
+        </Reference>
+    </ItemGroup>
+
 </Project>
