+++ src/Baballonia/Factories/CameraServiceFactory.cs
@@ -1,0 +1,6 @@
+﻿namespace Baballonia.Factories;
+
+public class CameraServiceFactory
+{
+    
+}
