+++ src/Baballonia/Services/Inference/CameraController.cs
@@ -67,16 +67,10 @@ public class CameraController : IDisposable
     // Tracking mode property
     private readonly StyledProperty<bool> _isTrackingModeProperty;
 
-    // MJPEG Streaming
-    private HttpListener _httpListener;
-    private bool _isStreaming;
-    private CancellationTokenSource _streamingCancellationTokenSource;
-    private readonly object _streamLock = new();
-    private byte[] _currentJpegFrame;
-    private int _mjpegPort = 8080;
-    private readonly string _mjpegBoundary = "mjpegstream";
-
     private (int, int) CameraSize { get; set; } = (0, 0);
+
+    private MjpegStreamingService _mjpegStreamingService;
+    private CameraService _cameraService;
 
     public CameraController(
         HomePageView view,
@@ -117,6 +111,9 @@ public class CameraController : IDisposable
         _flipYSettingKey = flipYSettingKey;
         _isTrackingModeProperty = isTrackingModeProperty;
 
+        _mjpegStreamingService = new MjpegStreamingService();
+        _cameraService = new CameraService(_localSettingsService, _inferenceService, camera);
+
         // Set up event handlers
         _mouthWindow.PointerPressed += OnPointerPressed;
         _mouthWindow.PointerMoved += OnPointerMoved;
@@ -128,9 +125,6 @@ public class CameraController : IDisposable
 
         // Set initial mode
         _view.SetValue(_isTrackingModeProperty, true);
-
-        // Initialize MJPEG streaming
-        _currentJpegFrame = [];
     }
 
     public async Task UpdateImage()
@@ -271,16 +265,12 @@ public class CameraController : IDisposable
 
     public void StartCamera(string cameraAddress)
     {
-        if (!string.IsNullOrEmpty(cameraAddress))
-        {
-            StopCamera();
-            _inferenceService.SetupInference(_camera, cameraAddress);
-        }
+        _cameraService.StartCamera(cameraAddress);
     }
 
     public void StopCamera()
     {
-        _inferenceService.Shutdown(_camera);
+        _cameraService.StopCamera();
     }
 
     public void SetTrackingMode()
@@ -416,273 +406,24 @@ public class CameraController : IDisposable
     /// <param name="port">Port to listen on (default: 8080)</param>
     public void StartMjpegStreaming(int port = 8080)
     {
-        if (_isStreaming)
-            return;
-
-        _mjpegPort = port;
-        _isStreaming = true;
-        _streamingCancellationTokenSource = new CancellationTokenSource();
-
-        try
-        {
-            _httpListener = new HttpListener();
-            _httpListener.Prefixes.Add($"http://localhost:{_mjpegPort}/");
-            _httpListener.Start();
-
-            Task.Run(() => HandleHttpRequests(_streamingCancellationTokenSource.Token));
-        }
-        catch (Exception)
-        {
-            _isStreaming = false;
-        }
-    }
-
-    /// <summary>
-    /// Stop the MJPEG streaming server
-    /// </summary>
-    public void StopMjpegStreaming()
-    {
-        if (!_isStreaming)
-            return;
-
-        _isStreaming = false;
-        _streamingCancellationTokenSource?.Cancel();
-
-        try
-        {
-            _httpListener?.Stop();
-            _httpListener?.Close();
-        }
-        catch (Exception)
-        {
-            // ignored
-        }
-    }
-
-    private async Task HandleHttpRequests(CancellationToken cancellationToken)
-    {
-        while (_isStreaming && !cancellationToken.IsCancellationRequested)
-        {
-            try
-            {
-                var context = await _httpListener.GetContextAsync();
-
-                if (cancellationToken.IsCancellationRequested)
-                    break;
-
-                string requestPath = context.Request.Url!.AbsolutePath.ToLowerInvariant();
-
-#pragma warning disable CS4014 // Awaiting prevents the calibration app from receiving MJPEG frames
-                if (requestPath == "/mjpeg")
-                {
-                    // Handle MJPEG stream request
-                    Task.Run(() => HandleMjpegRequest(context, cancellationToken));
-                }
-                else if (requestPath == "/snapshot" || requestPath == "/jpeg")
-                {
-                    // Handle single JPEG snapshot request
-                    Task.Run(() => HandleSnapshotRequest(context));
-                }
-                else
-                {
-                    // Handle other requests (like a simple status page)
-                    HandleDefaultRequest(context);
-                }
-#pragma warning restore CS4014 // Awaiting prevents the calibration app from receiving MJPEG frames
-            }
-            catch (Exception)
-            {
-                // ignored
-            }
-        }
-    }
-
-    private async Task HandleMjpegRequest(HttpListenerContext context, CancellationToken cancellationToken)
-    {
-        // Get the underlying TCP connection
-        var response = context.Response;
-
-        try
-        {
-            // Write the HTTP response headers manually
-            string responseHeaders =
-                "HTTP/1.1 200 OK\r\n" +
-                $"Content-Type: multipart/x-mixed-replace; boundary={_mjpegBoundary}\r\n" +
-                "Cache-Control: no-cache, no-store, must-revalidate\r\n" +
-                "Pragma: no-cache\r\n" +
-                "Expires: 0\r\n" +
-                "Connection: close\r\n\r\n";
-
-            byte[] headerBytes = Encoding.ASCII.GetBytes(responseHeaders);
-
-            // Get the raw output stream
-            using var outputStream = response.OutputStream;
-            await outputStream.WriteAsync(headerBytes, 0, headerBytes.Length, cancellationToken);
-
-            // Initial boundary
-            string initialBoundary = $"--{_mjpegBoundary}\r\n";
-            byte[] boundaryBytes = Encoding.ASCII.GetBytes(initialBoundary);
-            await outputStream.WriteAsync(boundaryBytes, 0, boundaryBytes.Length, cancellationToken);
-
-            // Stream frames
-            while (_isStreaming && !cancellationToken.IsCancellationRequested)
-            {
-                if (_currentJpegFrame == null || _currentJpegFrame.Length == 0)
-                {
-                    await Task.Delay(33, cancellationToken);
-                    continue;
-                }
-
-                // Copy to avoid race conditions
-                byte[] frameData = _currentJpegFrame;
-
-                try
-                {
-                    // Frame headers
-                    string frameHeader = "Content-Type: image/jpeg\r\n" +
-                                         $"Content-Length: {frameData.Length}\r\n" +
-                                         $"X-Timestamp: {DateTimeOffset.UtcNow.ToUnixTimeMilliseconds()}\r\n\r\n";
-                    byte[] frameHeaderBytes = Encoding.ASCII.GetBytes(frameHeader);
-
-                    await outputStream.WriteAsync(frameHeaderBytes, 0, frameHeaderBytes.Length, cancellationToken);
-                    await outputStream.WriteAsync(frameData, 0, frameData.Length, cancellationToken);
-
-                    // Next boundary
-                    string nextBoundary = $"\r\n--{_mjpegBoundary}\r\n";
-                    byte[] nextBoundaryBytes = Encoding.ASCII.GetBytes(nextBoundary);
-                    await outputStream.WriteAsync(nextBoundaryBytes, 0, nextBoundaryBytes.Length, cancellationToken);
-
-                    await outputStream.FlushAsync(cancellationToken);
-                }
-                catch (IOException)
-                {
-                    break;
-                }
-
-                await Task.Delay(33, cancellationToken);
-            }
-        }
-        catch (Exception)
-        {
-            // ignored
-        }
-        finally
-        {
-            try { response.Abort(); }
-            catch
-            {
-                // ignored
-            }
-        }
-    }
-
-    private void HandleSnapshotRequest(HttpListenerContext context)
-    {
-        HttpListenerResponse response = context.Response;
-
-        try
-        {
-            response.ContentType = "image/jpeg";
-            response.Headers.Add("Cache-Control", "no-cache, no-store, must-revalidate");
-            response.Headers.Add("Pragma", "no-cache");
-            response.Headers.Add("Expires", "0");
-
-            // Copy to avoid race conditions
-            byte[] frameData = _currentJpegFrame;
-
-            if (frameData.Length > 0)
-            {
-                //response.ContentLength64 = frameData.Length;
-                response.OutputStream.Write(frameData, 0, frameData.Length);
-            }
-            else
-            {
-                response.StatusCode = 503; // Service Unavailable
-                byte[] errorMsg = Encoding.UTF8.GetBytes("No image available");
-                //response.ContentLength64 = errorMsg.Length;
-                response.OutputStream.Write(errorMsg, 0, errorMsg.Length);
-            }
-        }
-        catch (Exception)
-        {
-            // ignored
-        }
-        finally
-        {
-            response.Close();
-        }
-    }
-
-    private void HandleDefaultRequest(HttpListenerContext context)
-    {
-        HttpListenerResponse response = context.Response;
-
-        try
-        {
-            string html = $@"
-                <!DOCTYPE html>
-                <html>
-                <head>
-                    <title>Camera Stream</title>
-                    <style>
-                        body {{ font-family: Arial, sans-serif; margin: 20px; }}
-                        h1 {{ color: #333; }}
-                        .stream-container {{ margin-top: 20px; }}
-                        img {{ max-width: 100%; border: 1px solid #ccc; }}
-                    </style>
-                </head>
-                <body>
-                    <h1>Camera Stream</h1>
-                    <div class='stream-container'>
-                        <h2>Live Stream</h2>
-                        <img src='/mjpeg' alt='MJPEG Stream'>
-                    </div>
-                    <div class='stream-container'>
-                        <h2>Static Image</h2>
-                        <img src='/snapshot' alt='Snapshot'>
-                    </div>
-                </body>
-                </html>";
-
-            byte[] buffer = Encoding.UTF8.GetBytes(html);
-            response.ContentType = "text/html";
-            //response.ContentLength64 = buffer.Length;
-            response.OutputStream.Write(buffer, 0, buffer.Length);
-        }
-        catch (Exception)
-        {
-            // ignored
-        }
-        finally
-        {
-            response.Close();
-        }
+        _mjpegStreamingService.StartStreaming(port);
     }
 
     private void UpdateMjpegFrame(Mat mat)
     {
-        if (_bitmap == null || !_isStreaming)
-            return;
+        _mjpegStreamingService.UpdateMjpegFrame(mat);
+    }
 
-        try
-        {
-            // Update the current frame
-            lock (_streamLock)
-            {
-                _currentJpegFrame = mat.ToBytes(ext: ".jpg"); // Cv2.Imencode
-            }
-        }
-        catch (Exception)
-        {
-            // ignored
-        }
+    public void StopMjpegStreaming()
+    {
+        _mjpegStreamingService.StopStreaming();
     }
 
     #endregion
 
     public void Dispose()
     {
-        StopMjpegStreaming();
+        _mjpegStreamingService.Dispose();
         StopCamera();
 
         _mouthWindow.PointerPressed -= OnPointerPressed;
@@ -690,6 +431,5 @@ public class CameraController : IDisposable
         _mouthWindow.PointerReleased -= OnPointerReleased;
 
         _bitmap?.Dispose();
-        _streamingCancellationTokenSource?.Dispose();
     }
 }
